# SwaggerOffLine.yaml
openapi: 3.1.0
info:
  title: SpecRadar
  description: "API de inteligência competitiva automotiva - padroniza fichas técnicas\
    \ de veículos a partir de atributos livres, usando um catálogo canônico e um serviç\
    o de Machine Learning para mapear os atributos solicitados."
  contact:
    name: FIAP - Desafio Ford 01
    email: specradar@email.com.br
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearer-key: []
...
paths:
  /veiculos/{id}:
    get:
      tags:
      - Veículos
      summary: Detalha um veículo e suas especificações
      operationId: detalharVeiculo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoVeiculo"
    put:
      tags:
      - Veículos
      summary: Atualiza os dados cadastrais de um veículo
      operationId: atualizarVeiculo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoVeiculo"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoVeiculo"
    delete:
      tags:
      - Veículos
      summary: Exclui (logicamente) um veículo
      operationId: excluirVeiculo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /veiculos:
    get:
      tags:
      - Veículos
      summary: "Lista os veículos ativos, de forma paginada"
      operationId: listarVeiculos
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "marca,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosListagemVeiculo"
    post:
      tags:
      - Veículos
      summary: Cadastra um veículo com suas especificações técnicas
      operationId: cadastrarVeiculo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroVeiculo"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoVeiculo"
...
components:
  schemas:
    DadosAtualizacaoVeiculo:
      type: object
      properties:
        marca:
          type: string
          maxLength: 80
          minLength: 0
        modelo:
          type: string
          maxLength: 80
          minLength: 0
        versao:
          type: string
          maxLength: 80
          minLength: 0
        ano:
          type: integer
          format: int32
    DadosDetalhamentoEspecificacao:
      type: object
      properties:
        atributo:
          type: string
        valor:
          type: string
        fonte:
          type: string
    DadosDetalhamentoVeiculo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        marca:
          type: string
        modelo:
          type: string
        versao:
          type: string
        ano:
          type: integer
          format: int32
        ativo:
          type: boolean
        especificacoes:
          type: array
          items:
            $ref: "#/components/schemas/DadosDetalhamentoEspecificacao"
...
    DadosCadastroEspecificacao:
      type: object
      properties:
        atributo:
          type: string
          maxLength: 60
          minLength: 0
        valor:
          type: string
          maxLength: 255
          minLength: 0
        fonte:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - atributo
      - valor
    DadosCadastroVeiculo:
      type: object
      properties:
        marca:
          type: string
          maxLength: 80
          minLength: 0
        modelo:
          type: string
          maxLength: 80
          minLength: 0
        versao:
          type: string
          maxLength: 80
          minLength: 0
        ano:
          type: integer
          format: int32
        especificacoes:
          type: array
          items:
            $ref: "#/components/schemas/DadosCadastroEspecificacao"
      required:
      - ano
      - marca
      - modelo
      - versao
...
    DadosListagemVeiculo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        marca:
          type: string
        modelo:
          type: string
        versao:
          type: string
        ano:
          type: integer
          format: int32
    PageDadosListagemVeiculo:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemVeiculo"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT